requirements.txt
scikit-learn==1.6.1
biopython
streamlit
joblib==1.4.2